implementing useffect
